INGM-741 add error queue for errors
