+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- Extended error queue manager (`ServoErrorQueue`) to support standard drive errors (MOCO/COCO/SYSTEM).
+- Added standard error queue descriptors: `MOCO_ERROR_QUEUE`, `COCO_ERROR_QUEUE`, `SYSTEM_ERROR_QUEUE`.
+
+### Deprecated
+- `Errors.get_all_errors()` is deprecated. Use `ServoErrorQueue` for race-safe error reading.
+
 ## [0.10.1] - 2025-11-24
 ### Added
 - Add `is_active` function in SafetyFunctions class.
